@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.10+11
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -32,8 +32,8 @@ dependencies:
   url_launcher: ^6.0.3
   firebase_core: ^1.0.4
   firebase_analytics: ^8.0.1
-  firebase_crashlytics: ^1.0.0
-  admob_flutter: ^1.0.1
+  firebase_crashlytics: ^2.0.1
+  admob_flutter: ^2.0.0-nullsafety.1
   flex_color_scheme: ^2.1.1
   lottie: ^1.0.1
 
@@ -41,7 +41,7 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
