@@ -4,14 +4,14 @@ import 'package:flutter/material.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({
-    Key key,
+    Key? key,
     this.themeMode,
     this.onThemeModeChanged,
     this.flexSchemeData,
   }) : super(key: key);
-  final ThemeMode themeMode;
-  final ValueChanged<ThemeMode> onThemeModeChanged;
-  final FlexSchemeData flexSchemeData;
+  final ThemeMode? themeMode;
+  final ValueChanged<ThemeMode>? onThemeModeChanged;
+  final FlexSchemeData? flexSchemeData;
 
   @override
   Widget build(BuildContext context) {
@@ -30,9 +30,9 @@ class HomePage extends StatelessWidget {
             ),
             selectedLabelStyle: TextStyle(color: Colors.white, fontWeight: FontWeight.bold, fontSize: 16.0),
             unselectedLabelStyle: TextStyle(color: Colors.grey),
-            themeMode: themeMode,
-            onThemeModeChanged: onThemeModeChanged,
-            flexSchemeData: flexSchemeData,
+            themeMode: themeMode!,
+            onThemeModeChanged: onThemeModeChanged!,
+            flexSchemeData: flexSchemeData!,
           ),
         ),
       ],
