@@ -9,13 +9,13 @@ import 'package:firebase_analytics/firebase_analytics.dart';
 
 class TabViewThree extends StatefulWidget {
   TabViewThree({
-    Key key,
+    Key? key,
     this.analytics,
     //   this.observer,
   }) : super(key: key);
 
   // Firebase Analytics
-  final FirebaseAnalytics analytics;
+  final FirebaseAnalytics? analytics;
 //  final FirebaseAnalyticsObserver observer;
 
   @override
@@ -24,9 +24,9 @@ class TabViewThree extends StatefulWidget {
 class _State extends State<TabViewThree> {
   ThemeMode themeMode = ThemeMode.light;
   // Firebase Analytics
-  FirebaseAnalytics _analytics;
+  late FirebaseAnalytics _analytics;
 
-  _State(FirebaseAnalytics analytics);
+  _State(FirebaseAnalytics? analytics);
 
   @override
   void initState() {
@@ -54,7 +54,7 @@ class _State extends State<TabViewThree> {
                   if (cart.count2 == 1 &&
                       arrays3[index].isFav == false) {
                     for (int i = 0;
-                    i < arrays3.length;
+                   i < arrays3.length;
                     i++) {
                       arrays3[i].player.pause();
                       cart.remove2(arrays3[i]);
@@ -65,7 +65,7 @@ class _State extends State<TabViewThree> {
                   if (cart.count2 <= 1 &&
                       arrays3[index].isFav == false) {
                     arrays3[index].player.open(
-                        Audio(arrays3[index].sounds),
+                        Audio(arrays3[index].sounds!),
                         volume: 0.5,
                         loopMode: LoopMode.single);
                     cart.add2(arrays3[index]);
@@ -85,9 +85,9 @@ class _State extends State<TabViewThree> {
                     foregroundServiceStop();
                   }
                   // Click_events - if isFav is true
-                 if (arrays3[index].isFav) {
+                 if (arrays3[index].isFav!) {
                     await _analytics.logEvent(
-                      name: arrays3[index].events,
+                      name: arrays3[index].events!,
                     );
                   }
                 },
@@ -98,16 +98,16 @@ class _State extends State<TabViewThree> {
                       width: 60,
                       //  height: 50.0,
                       image: AssetImage(
-                          arrays3[index].isFav
-                              ? arrays3[index].picOn
-                              : arrays3[index].picOff),
+                          arrays3[index].isFav!
+                              ? arrays3[index].picOn!
+                              : arrays3[index].picOff!),
                     ),
                     Padding(
                         padding:
                         EdgeInsets.only(top: 10)),
-                    arrays3[index].isFav ? AnimatedOpacity(
+                    arrays3[index].isFav! ? AnimatedOpacity(
                       duration: Duration(milliseconds: 700),
-                      opacity: arrays3[index].isFav
+                      opacity: arrays3[index].isFav!
                           ? arrays3[index].opacityOn
                           : arrays3[index].opacityOff,
                       child: PlayerBuilder.volume(
@@ -128,7 +128,7 @@ class _State extends State<TabViewThree> {
                                 });
                           }),
                     ) :  Text(
-                      arrays3[index].title,
+                      arrays3[index].title!,
                       style: TextStyle(fontSize: 13,
                       color: Colors.white),
                       textAlign: TextAlign.center,
